This id the circuit for the question 1
This file is a logisim file. This file cannot be opened directly please install this file and then open the file from the logi sim app.
In this file a right rotating circuit is made with the concept of the universal shift register which is implemented using a 2:1 mux. Here the clock signal is the number of times the number has to be rotated towards the right. There is the option of parallel loading the registers by the additional 2:1mux. 
For a n bit number we may require a 3n number of 2:1 mux. 
3 -  2:1 mux for one d flip flop
1 - 2:1 mux for the option for loading and rotating